+++ src/lib.rs
@@ -170,7 +170,7 @@ mod test_utils;
 
 use properties::PropertyInstance;
 
-pub use asset::{EffectAsset, MotionIntegration, SimulationCondition};
+pub use asset::{AlphaMode, EffectAsset, MotionIntegration, SimulationCondition};
 pub use attributes::*;
 pub use bundle::ParticleEffectBundle;
 pub use gradient::{Gradient, GradientKey};
@@ -178,7 +178,7 @@ pub use graph::*;
 pub use modifier::*;
 pub use plugin::HanabiPlugin;
 pub use properties::{Property, PropertyLayout};
-pub use render::{EffectSystems, ShaderCache};
+pub use render::{EffectSystems, LayoutFlags, ShaderCache};
 pub use spawn::{tick_spawners, CpuValue, EffectSpawner, Random, Spawner};
 
 #[allow(missing_docs)]
@@ -611,10 +611,8 @@ pub struct CompiledParticleEffect {
     /// 2D layer for the effect instance.
     #[cfg(feature = "2d")]
     z_layer_2d: FloatOrd,
-    /// Is the particle size in screen-space logical pixels?
-    screen_space_size: bool,
-    /// Is the effect simulated in local space?
-    local_space_simulation: bool,
+    /// Layout flags.
+    layout_flags: LayoutFlags,
 }
 
 impl Default for CompiledParticleEffect {
@@ -630,8 +628,7 @@ impl Default for CompiledParticleEffect {
             particle_texture: None,
             #[cfg(feature = "2d")]
             z_layer_2d: FloatOrd(0.0),
-            screen_space_size: false,
-            local_space_simulation: false,
+            layout_flags: LayoutFlags::NONE,
         }
     }
 }
@@ -857,15 +854,40 @@ impl CompiledParticleEffect {
             }
         }
 
-        self.local_space_simulation = asset.simulation_space == SimulationSpace::Local;
-
         // Generate the shader code for the render shader
-        let mut render_context = RenderContext::default();
+        let mut render_context = RenderContext::new(&mut module, &property_layout);
         for m in asset.render_modifiers() {
             m.apply_render(&mut render_context);
         }
+        let alpha_cutoff_code = if let AlphaMode::Mask(cutoff) = &asset.alpha_mode {
+            render_context.eval(*cutoff).unwrap_or_else(|err| {
+                error!(
+                    "Failed to evaluate the expression for AlphaMode::Mask, error: {:?}",
+                    err
+                );
 
-        self.screen_space_size = render_context.screen_space_size;
+                // In Debug, show everything to help diagnosing
+                #[cfg(debug_assertions)]
+                return 1_f32.to_wgsl_string();
+
+                // In Release, hide everything with an error
+                #[cfg(not(debug_assertions))]
+                return 0_f32.to_wgsl_string();
+            })
+        } else {
+            String::new()
+        };
+
+        self.layout_flags = LayoutFlags::NONE;
+        if asset.simulation_space == SimulationSpace::Local {
+            self.layout_flags |= LayoutFlags::LOCAL_SPACE_SIMULATION;
+        }
+        if let AlphaMode::Mask(_) = &asset.alpha_mode {
+            self.layout_flags |= LayoutFlags::USE_ALPHA_MASK;
+        }
+        if render_context.screen_space_size {
+            self.layout_flags |= LayoutFlags::SCREEN_SPACE_SIZE;
+        }
 
         // Configure aging code
         let has_age = present_attributes.contains(&Attribute::AGE);
@@ -954,18 +976,18 @@ impl CompiledParticleEffect {
             .replace(
                 "{{SIMULATION_SPACE_TRANSFORM_PARTICLE}}",
                 sim_space_transform_code,
-            );
+            )
+            .replace("{{ALPHA_CUTOFF}}", &alpha_cutoff_code);
         let render_shader = shader_cache.get_or_insert(&asset.name, &render_shader_source, shaders);
         trace!("Configured render shader:\n{}", render_shader_source);
 
         trace!(
-            "tick_spawners: init_shader={:?} update_shader={:?} render_shader={:?} has_image={} screen_space_size={} local_space_simulation={}",
+            "tick_spawners: init_shader={:?} update_shader={:?} render_shader={:?} has_image={} layout_flags={:?}",
             init_shader,
             update_shader,
             render_shader,
             render_context.particle_texture.is_some(),
-            self.screen_space_size,
-            self.local_space_simulation
+            self.layout_flags,
         );
 
         // TODO - Replace with Option<ConfiguredShader { handle: Handle<Shader>, hash: